ui_area: ui areas now handle key and pointer events - key_input: no real distinction between button and key events - now responding to pointer button release events - if active area does not handle key event then its passed to parent

+++ include/ui/key_action.h
@@ -5,8 +5,8 @@
 
   typedef void (*ActionFn)(const short souce_id, const short destination_id);
 
-  short Key_Action_create(const short key, ActionFn action, const char* help);
-  short Key_Action_key(const short key_action_id);
+  short Key_Action_create(const int key, ActionFn action, const char* help);
+  int Key_Action_key(const short key_action_id);
   void Key_Action_execute(const short key_action_id, const short source_id,  const short destination_id);
   void Key_Action_print_help(void);
 #endif
